requirements.txt
Flask
numpy
pandas
scikit-learn
matplotlib
fpdf2
